requirements.txt
llama-cpp-python>=0.2.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
pytest>=7.0.0
authlib>=1.2.0
itsdangerous>=2.1.0
PyJWT>=2.8.0
httpx>=0.25.0
bcrypt>=4.0.0
